refactor(graph_view): extract row-folding/viewport module (#79)
Pure code move: visible_nodes, RenderRow, visible_rows, fold_rows,
fold_rows_windowed, merge_connector_cells, adjacent_cells, and
visible_row_window (plus the window_covers_* and connector-folding test
groups) move into a new rows submodule. visible_nodes, visible_rows, and
RenderRow are re-exported so the graph_view::* contract is preserved; the
internally-used helpers are pub(super). fold_rows stays pub(super) so a
tracing test still living in mod.rs keeps compiling. No behavior changes.
